Add LoadingIndicator Component and MutationalSignaturesUtils
