+++ src/pages/patientView/PatientViewPage.tsx
@@ -251,9 +251,12 @@ export default class PatientViewPage extends React.Component<IPatientViewPagePro
                                 )
                             }
                         </span>
+                        {patientViewPageStore.hasMutationalSignatureData.isComplete &&
+                        <LoadingIndicator isLoading={patientViewPageStore.mutationalSignatureData.isPending && patientViewPageStore.mutationalSignatureMetaData.isPending}/>}
 
-                        {patientViewPageStore.clinicalDataGroupedBySample.isComplete && patientViewPageStore.mutationalSignatureData.isComplete
-                        && patientViewPageStore.mutationalSignatureMetaData.isComplete && patientViewPageStore.hasMutationalSignatureData.isComplete &&
+                        {patientViewPageStore.hasMutationalSignatureData.isComplete &&
+                        patientViewPageStore.clinicalDataGroupedBySample.isComplete && patientViewPageStore.mutationalSignatureData.isComplete &&
+                        patientViewPageStore.mutationalSignatureMetaData.isComplete &&
                         (<SignificantMutationalSignatures data={patientViewPageStore.mutationalSignatureData.result}
                         metadata={patientViewPageStore.mutationalSignatureMetaData.result} uniqueSampleKey={sample.id}/>)}
 
@@ -507,10 +510,10 @@ export default class PatientViewPage extends React.Component<IPatientViewPagePro
                         </div>
                     </MSKTab>
 
-                    <MSKTab key={6} id="mutationalSignatureTab" linkText="Mutational Signature Data">
-
+                    <MSKTab key={6} id="mutationalSignatureTab" linkText="Mutational Signature Data" hide={!patientViewPageStore.hasMutationalSignatureData.isComplete}>
                         <div className="clearfix">
                             <FeatureTitle title="Mutational Signatures" isLoading={ patientViewPageStore.clinicalDataGroupedBySample.isPending } className="pull-left" />
+                            <LoadingIndicator isLoading={patientViewPageStore.mutationalSignatureData.isPending}/>
                             {  (patientViewPageStore.clinicalDataGroupedBySample.isComplete && patientViewPageStore.mutationalSignatureData.isComplete) && (
                                 <ClinicalInformationMutationalSignatureTable data={patientViewPageStore.mutationalSignatureData.result} showTitleBar={true}/>
                             )
